+++ src/auth/auth.controller.ts
@@ -5,11 +5,13 @@ import { ConfigService } from '@nestjs/config';
 import { Response } from 'express';
 import { AuthService } from './services/auth.service';
 import { JwtGuard } from './jwt/jwt.guard';
+import { CookieService } from './services/cookie.service';
 
 @Controller('auth')
 export class AuthController {
   constructor(
     private readonly authService: AuthService,
+    private readonly cookieService: CookieService,
     private readonly configService: ConfigService,
   ) {}
 
@@ -25,38 +27,13 @@ export class AuthController {
     const { accessToken, refreshToken } =
       await this.authService.googleLogin(user);
 
-    // *쿠키 설정 옵션 분리예정
-    // 쿠키와 리다이렉트의 설정은 클라이언트가 원하는 곳으로 지정해준다.
-    // res.cookie 의 domain부분과 res.redirect의 url부분의 도메인을 일치시켜야 한다.
-    const cookieOptions = {
-      httpOnly: true,
-      secure: true,
-      domain: this.configService.get<string>('COOKIE_DOMAIN'),
-      sameSite: 'none' as 'none', // none 타입으로 지정해줘야 함.
-      maxAge: this.configService.get<number>('COOKIE_EXPIRATION'),
-    };
-
-    const cookies = [
-      {
-        name: 'accessToken',
-        value: accessToken,
-        options: { ...cookieOptions },
-      },
-      {
-        name: 'refreshToken',
-        value: refreshToken,
-        options: { ...cookieOptions }, // path 추가 예정
-      },
-    ];
-
-    for (const cookie of cookies) {
-      response.cookie(cookie.name, cookie.value, cookie.options);
-    }
-
-    response.redirect(this.configService.get<string>('CLIENT_MAIN_PAGE_URL'));
+    await this.cookieService.cookieResponse(
+      response,
+      accessToken,
+      refreshToken,
+    );
   }
 
-  @Post('/logout')
   @Get('/verify')
   @UseGuards(JwtGuard)
   tokenVerify(@User() user: any) {
