+++ src/auth/services/cookie.service.ts
@@ -1,0 +1,45 @@
+import { Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+
+@Injectable()
+export class CookieService {
+  constructor(private configService: ConfigService) {}
+
+  async cookieResponse(
+    response: any,
+    accessToken: string,
+    refreshToken: string,
+  ) {
+    // 쿠키와 리다이렉트의 설정은 클라이언트가 원하는 곳으로 지정해준다.
+    // res.cookie 의 domain부분과 res.redirect의 url부분의 도메인을 일치시켜야 한다.
+    const cookieOptions = {
+      httpOnly: true,
+      secure: true,
+      domain: this.configService.get<string>('COOKIE_DOMAIN'),
+      sameSite: 'none' as 'none', // none 타입으로 지정해줘야 함.
+      maxAge: this.configService.get<number>('COOKIE_EXPIRATION'),
+    };
+
+    // access 와 refresh 토큰 옵션 설정
+    const cookies = [
+      {
+        name: 'accessToken',
+        value: accessToken,
+        options: { ...cookieOptions },
+      },
+      {
+        name: 'refreshToken',
+        value: refreshToken,
+        options: { ...cookieOptions }, // path 추가 예정
+      },
+    ];
+
+    // 쿠키 설정정
+    for (const cookie of cookies) {
+      response.cookie(cookie.name, cookie.value, cookie.options);
+    }
+
+    // 메인페이지로 리다이렉트
+    response.redirect(this.configService.get<string>('CLIENT_MAIN_PAGE_URL'));
+  }
+}
